# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  models.APIResponse:
    properties:
      code:
        description: 错误码
        type: integer
      data:
        description: 数据
      errors:
        description: 字段错误列表
        items:
          $ref: '#/definitions/models.FieldError'
        type: array
      message:
        description: 消息
        type: string
    type: object
...
  models.AdminStatsResponse:
    properties:
      active_users:
        type: integer
      admin_users:
        type: integer
      member_users:
        type: integer
      new_users_month:
        type: integer
      new_users_today:
        type: integer
      new_users_week:
        type: integer
      total_users:
        type: integer
    type: object
  models.AdminUserInfo:
    properties:
      created_at:
        type: string
      default_role:
        $ref: '#/definitions/models.UserRole'
      email:
        type: string
      email_verified:
        type: boolean
      id:
        type: string
      last_login_at:
        type: string
      login_count:
        type: integer
      membership_expiry:
        type: string
      roles:
        items:
          $ref: '#/definitions/models.RoleInfo'
        type: array
      status:
        $ref: '#/definitions/models.UserStatus'
      updated_at:
        type: string
      username:
        type: string
    type: object
  models.AssignRoleRequest:
    properties:
      expires_at:
        type: string
      role_id:
        type: string
      user_id:
        type: string
    required:
    - role_id
    - user_id
    type: object
...
  models.FieldError:
    properties:
      field:
        description: 字段名
        type: string
      message:
        description: 错误信息
        type: string
    type: object
...
  models.PaginatedResponse:
    properties:
      data: {}
      pagination:
        $ref: '#/definitions/models.Pagination'
    type: object
  models.Pagination:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  models.RoleInfo:
    properties:
      description:
        type: string
      display_name:
        type: string
      expires_at:
        type: string
      id:
        type: string
      level:
        type: integer
      name:
        $ref: '#/definitions/models.UserRole'
    type: object
...
  models.UpdateUserStatusRequest:
    properties:
      reason:
        maxLength: 255
        type: string
      status:
        allOf:
        - $ref: '#/definitions/models.UserStatus'
        enum:
        - active
        - suspended
        - deleted
    required:
    - status
    type: object
...
  models.UserRole:
    enum:
    - regular
    - member
    - admin
    - super_admin
    type: string
    x-enum-comments:
      UserRoleAdmin: 管理员
      UserRoleMember: 会员用户
      UserRoleRegular: 普通用户
      UserRoleSuperAdmin: 超级管理员
    x-enum-descriptions:
    - 普通用户
    - 会员用户
    - 管理员
    - 超级管理员
    x-enum-varnames:
    - UserRoleRegular
    - UserRoleMember
    - UserRoleAdmin
    - UserRoleSuperAdmin
...
  models.UserStatus:
    enum:
    - active
    - suspended
    - deleted
    type: string
    x-enum-varnames:
    - UserStatusActive
    - UserStatusSuspended
    - UserStatusDeleted
...
host: localhost:8080
info:
  contact:
    email: support@oraura.app
    name: OrAura API Support
    url: https://oraura.app/support
  description: OrAura spiritual divination application backend service with user management
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://oraura.app/terms
  title: OrAura Backend API
  version: "1.0"
paths:
  /admin/logs/login:
    get:
      description: 管理员查看系统登录日志
      parameters:
      - in: query
        name: end_date
        type: string
      - in: query
        name: ip_address
        type: string
      - in: query
        minimum: 1
        name: page
        type: integer
      - in: query
        maximum: 100
        minimum: 1
        name: per_page
        type: integer
      - in: query
        name: sort_desc
        type: boolean
      - in: query
        name: start_date
        type: string
      - in: query
        name: success
        type: boolean
      - in: query
        name: user_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.PaginatedResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - BearerAuth: []
      summary: 获取登录日志
      tags:
      - 管理员
  /admin/stats:
    get:
      description: 获取用户数量、活跃用户等统计信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.AdminStatsResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - BearerAuth: []
      summary: 获取管理员仪表板统计信息
      tags:
      - 管理员
...
  /admin/users:
    get:
      description: 管理员查看所有用户列表，支持筛选和分页
      parameters:
      - in: query
        minimum: 1
        name: page
        type: integer
      - in: query
        maximum: 100
        minimum: 1
        name: per_page
        type: integer
      - enum:
        - regular
        - member
        - admin
        - super_admin
        in: query
        name: role
        type: string
        x-enum-comments:
          UserRoleAdmin: 管理员
          UserRoleMember: 会员用户
          UserRoleRegular: 普通用户
          UserRoleSuperAdmin: 超级管理员
        x-enum-descriptions:
        - 普通用户
        - 会员用户
        - 管理员
        - 超级管理员
        x-enum-varnames:
        - UserRoleRegular
        - UserRoleMember
        - UserRoleAdmin
        - UserRoleSuperAdmin
      - in: query
        maxLength: 100
        name: search
        type: string
      - enum:
        - created_at
        - updated_at
        - login_count
        - last_login_at
        in: query
        name: sort_by
        type: string
      - in: query
        name: sort_desc
        type: boolean
      - enum:
        - active
        - suspended
        - deleted
        in: query
        name: status
        type: string
        x-enum-varnames:
        - UserStatusActive
        - UserStatusSuspended
        - UserStatusDeleted
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.PaginatedResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - BearerAuth: []
      summary: 获取用户列表
      tags:
      - 管理员
  /admin/users/{user_id}:
    get:
      description: 管理员查看用户详细信息
      parameters:
      - description: 用户ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.AdminUserInfo'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - BearerAuth: []
      summary: 获取用户详情
      tags:
      - 管理员
  /admin/users/{user_id}/roles:
    post:
      consumes:
      - application/json
      description: 管理员为用户分配角色
      parameters:
      - description: 用户ID
        in: path
        name: user_id
        required: true
        type: string
      - description: 角色分配请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.AssignRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - BearerAuth: []
      summary: 分配角色给用户
      tags:
      - 管理员
  /admin/users/{user_id}/roles/{role_id}:
    delete:
      description: 管理员撤销用户的角色
      parameters:
      - description: 用户ID
        in: path
        name: user_id
        required: true
        type: string
      - description: 角色ID
        in: path
        name: role_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - BearerAuth: []
      summary: 撤销用户角色
      tags:
      - 管理员
  /admin/users/{user_id}/status:
    put:
      consumes:
      - application/json
      description: 管理员更新用户状态（激活/暂停/删除）
      parameters:
      - description: 用户ID
        in: path
        name: user_id
        required: true
        type: string
      - description: 状态更新请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateUserStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - BearerAuth: []
      summary: 更新用户状态
      tags:
      - 管理员
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"